AutoService/requirements.txt

asgiref==3.8.1
asttokens==2.4.1



colorama==0.4.6
comm==0.2.0
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Django==5.2.1
django-allauth==65.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
psycopg2-binary==2.9.10

executing==2.0.1
fastjsonschema==2.19.0
ipykernel==6.26.0
ipython==8.12.3

python-dotenv==1.0.1
PyJWT==2.9.0

python-json-logger==2.0.7



PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
requests==2.31.0



tzdata==2023.3

